files and folders of sass restructured to get more order, file variables.scss exported to the others files of scss
