+++ src/App.js
@@ -1,9 +1,8 @@
 import React, { Fragment } from 'react';
-import '../src/Styles/App.scss'
+import '../src/assets/styles/App.scss';
 import NavBar from './components/Menu/Navbar';
 import LinksRRSS from './components/Menu/LinksRRSS';
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom"
-import '../src/Styles/Slider.scss'
 import Slider from './components/slider/Slider';
 
 function App() {
