+++ src/assets/styles/App.scss
@@ -1,0 +1,20 @@
+@import './Variables.scss'; 
+
+*{
+    padding: 0;
+    margin: 0;
+    font-family:'Montserrat', sans-serif; 
+   
+}
+
+
+
+ 
+
+
+
+
+
+
+
+
