+++ src/assets/styles/Variables.scss
@@ -1,0 +1,10 @@
+$firstColor: #FF5920;
+$secondColor:#00DBD3;
+$thirdColor:#9B51E0;
+$fourthColor:#6AEE67;
+$textColor: #828282;
+$textColor1:#333333;
+$textColor2:#F2C94C;
+$textColor3:#FFFFFF;
+
+
