+++ src/assets/styles/components/BotonDonar.scss
@@ -1,0 +1,22 @@
+@import '../Variables.scss'; 
+
+// boton donar
+.btnDonate{
+   text-decoration: none;
+   color: $textColor3;
+   font-size: 0.9em;
+
+}
+
+.containerBtnDonate{
+width: 145px;
+height: 32px;
+background-color: $firstColor;
+color: $textColor3;
+font-weight: bold;
+border-radius: 5px;
+margin-top: 20px;
+text-align: center;
+padding-top: 8px;
+
+}
