+++ src/assets/styles/components/LinksRRSS.scss
@@ -1,0 +1,18 @@
+@import '../Variables.scss'; 
+
+// estilos RRSS
+.boxContainerRRSS{
+    display: flex;
+    justify-content: space-between;
+    margin-top: 27px;
+    margin-right: 50px;
+    margin-left: 4vw;
+
+    
+}
+.containerRRSS {
+    //margin-top: 2vh;
+    font-size: 1.5rem;
+    margin-left: 10px;
+
+}
