+++ src/assets/styles/components/Logo.scss
@@ -1,0 +1,14 @@
+@import '../Variables.scss'; 
+
+// estilo logo
+.container-logo{
+    width: 4.3vw;
+    margin-top: 1.7vh;
+    margin-left: 3vh;
+    
+}
+
+.imgLogo {
+    width:100%; 
+   
+}
