+++ src/components/Menu/Navbar.jsx
@@ -3,7 +3,7 @@ import LogoHome from './Logo';
 import LinksNavbar from './LinksNavbar';
 import LinksRRSS from './LinksRRSS';
 import BtnDonate from './BotonDonar';
-
+import '../../assets/styles/components/Navbar.scss'; 
 
 const Navbar = () => {
     return (
