+++ src/components/Slider/Imgcomp.jsx
@@ -1,4 +1,5 @@
-import React from 'react';
+import React from 'react';  
+// import '../../assets/styles/components/Imgcomp.scss'; 
 
 function ImgComp({ src }) {
     let imgStyles ={
