+++ src/components/Slider/Slider.jsx
@@ -1,12 +1,13 @@
 import React,{useState}  from 'react';
-import ImgComp from './Imgcomp'
-import i1 from "./Pics/1.jpg";
-import i2 from "./Pics/2.jpg";
-import i3 from "./Pics/3.jpg";
+import ImgComp from './Imgcomp'; 
+import i1 from "../../assets/static/Pics/1.jpg";
+import i2 from "../../assets/static/Pics/2.jpg";
+import i3 from "../../assets/static/Pics/3.jpg";
 // npm install --save-dev @iconify/react @iconify/icons-bytesize
 import { Icon} from '@iconify/react';
 import chevronLeft from '@iconify/icons-bytesize/chevron-left';
 import chevronRight from '@iconify/icons-bytesize/chevron-right';
+import '../../assets/styles/components/Slider.scss'; 
 
 
 function Slider() {
@@ -35,7 +36,6 @@ function Slider() {
             <button id="goLeft" onClick={goLeft}><Icon icon={chevronLeft} color="#FFFFFF" width="3vh" /></button>
             <button id="goRight" onClick={goRight}><Icon icon={chevronRight} color="#FFFFFF" width="3vh" /></button>
     </div>
-
     )
 }
 export default Slider
