+++ src/components/Slider/SliderTxt.jsx
@@ -1,0 +1,14 @@
+import React from 'react'; 
+import BtnDonate from '../Menu/BotonDonar'; 
+
+const SliderTxt = () => {
+    return (
+        <div>
+            <h1>700mil personas padecen inseguridad alimentaria severa en Chile</h1>
+            <h2>Ayúdanos a combatir el hambre en nuestras comunidades</h2>
+            <BtnDonate />
+        </div>
+    )
+}
+
+export default SliderTxt
